Add updates before moving to ssh

+++ Assignment1/q2/seq_col_trs.c
@@ -6,25 +6,25 @@
 
 #define SIZE (10)
 
-float **L, *y, *x;
+double **L, *y, *x;
 int n;
 
-void InitializeInput(float** L, float* x, float* y)
+void InitializeInput(double** L, double* x, double* y)
 {
     for (int i = 0; i < n; i++)
     {
         for (int j = 0; j <= i; j++)
         {
-            L[i][j] = ((float)(random() % 100)) / 100.0 + 0.01;
+            L[i][j] = ((double)(random() % 100)) / 100.0 + 0.01;
         }
     }
     for (int i = 0; i < n; i++)
     {
-        y[i] = ((float)(random() % 100)) / 100.0 + 0.01;
+        y[i] = ((double)(random() % 100)) / 100.0 + 0.01;
     }
     for (int i = 0; i < n; i++)
     {
-        x[i] = ((float)(random() % 100)) / 100.0 + 0.01;
+        x[i] = ((double)(random() % 100)) / 100.0 + 0.01;
     }
 }
 
@@ -32,28 +32,28 @@ int main(int argc, char *argv[])
 {
     n = strtol(argv[1], NULL, 10);
     // Input taking!!
-    L = (float**)malloc( (n)*sizeof(float*) );
+    L = (double**)malloc( (n)*sizeof(double*) );
     assert(L != NULL);
     for (int i = 0; i < n; i++)
     {
-        L[i] = (float*)malloc((i + 1)*sizeof(float));
+        L[i] = (double*)malloc((i + 1)*sizeof(double));
         assert(L[i] != NULL);
 
         // for (int j = 0; j <= i; j++)
         // {
-        //     scanf("%f", &L[i][j]);
+            // scanf("%f", &L[i][j]);
         // }
 
     }
-    y = (float*)malloc((n) * sizeof(float));
-    x = (float*)malloc((n) * sizeof(float));
+    y = (double*)malloc((n) * sizeof(double));
+    x = (double*)malloc((n) * sizeof(double));
     assert(y != NULL);
     assert(x != NULL);
     InitializeInput(L, x, y);
     // for (int i = 0; i < n; i++)
     // {
     //     scanf("%f", &y[i]);
-    //     x[i] = ((float)(random() % 100)) / 100;
+    //     x[i] = ((double)(random() % 100)) / 100;
     // }
 
     // Solving
